refactor: `Route` internal module
